moves destination adapters to dlt.destination.adapters
